MVP-82 [refactor] factor out server data & `upload` func
