+++ voyager/Server.swift
@@ -5,52 +5,7 @@
 //  Created by 지우석 on 2022/07/27.
 //
 
-import Foundation
-import CoreVideo
 import Alamofire
-import CoreImage
-import UIKit
-
-struct ServerImageUploadData {
-    static var compressionQuality = 0.5
-    static var targetSize = CGSize(width: 160, height: 320)
-    
-    var filename = "test.jpg"
-    var sequenceNo = 0
-    var image: CVPixelBuffer
-    var imageData: Data? {
-        let ciImage = CIImage(cvPixelBuffer: image).oriented(.right)
-        let scaledUIImage = UIImage(ciImage: ciImage).scalePreservingAspectRatio(targetSize: ServerImageUploadData.targetSize)
-        
-        return scaledUIImage.jpegData(compressionQuality: ServerImageUploadData.compressionQuality)
-    }
-}
-
-struct ServerImageResponseRawData: Decodable {
-    let sequenceNo: Int
-}
-
-struct ServerYoloResponseRawData: Decodable {
-    let xmin: Double
-    let xmax: Double
-    let ymin: Double
-    let ymax: Double
-    
-    let classId: Int
-    let className: String
-    
-    enum CodingKeys: String, CodingKey {
-        case xmin
-        case xmax
-        case ymin
-        case ymax
-        
-        case classId = "class"
-        case className = "name"
-    }
-}
-
-typealias ServerGuideResponseRawData = String
 
 protocol ServerGuideDelegate: AnyObject {
     func alertGuide(guide: [ServerGuideResponseRawData])
@@ -90,22 +45,21 @@ class Server {
     
     func stop() {}
     
+    func getMultipartAppender(imgData: ServerImageUploadData) -> ((MultipartFormData) -> Void) {
+        return { data in
+            data.append(imgData.filename.data(using: .utf8)!, withName: "filename", mimeType: "text/plain")
+            data.append("\(imgData.sequenceNo)".data(using: .utf8)!, withName: "sequenceNo", mimeType: "text/plain")
+            data.append(imgData.imageData!, withName: "img", fileName: imgData.filename, mimeType: "image/jpg")
+        }
+    }
+    
     func send(imgData: ServerImageUploadData) {
         print("sending image to server")
         
-        let parameters: [String: Any] = [
-            "filename": imgData.filename,
-            "sequenceNo": imgData.sequenceNo
-        ]
-        
-        _ = AF.upload(multipartFormData: { multipartFormData in
-            for (key, value) in parameters {
-                multipartFormData.append("\(value)".data(using: .utf8)!, withName: key, mimeType: "text/plain")
-            }
-            
-            multipartFormData.append(imgData.imageData!, withName: "img", fileName: imgData.filename, mimeType: "image/jpg")
-                
-        }, to: serverImageUploadURI!).responseDecodable(of: [ServerGuideResponseRawData].self) { [weak self] response in
+        _ = AF.upload(
+            multipartFormData: getMultipartAppender(imgData: imgData),
+            to: serverImageUploadURI!
+        ).responseDecodable(of: [ServerGuideResponseRawData].self) { [weak self] response in
             print("response recieved: \(response)")
             if case .failure = response.result {
                 print(response.debugDescription)
